+++ molecule/config.py
@@ -31,6 +31,7 @@ from molecule import util
 from molecule.dependency import ansible_galaxy
 from molecule.dependency import gilt
 from molecule.driver import dockr
+from molecule.driver import ec2
 from molecule.driver import lxc
 from molecule.driver import lxd
 from molecule.driver import openstack
@@ -101,6 +102,8 @@ class Config(object):
 
         if driver_name == 'docker':
             driver = dockr.Dockr(self)
+        elif driver_name == 'ec2':
+            driver = ec2.Ec2(self)
         elif driver_name == 'lxc':
             driver = lxc.Lxc(self)
         elif driver_name == 'lxd':
@@ -331,6 +334,7 @@ def molecule_file(path):
 def molecule_drivers():
     return [
         dockr.Dockr(None).name,
+        ec2.Ec2(None).name,
         lxc.Lxc(None).name,
         lxd.Lxd(None).name,
         openstack.Openstack(None).name,